requirements.txt
python-dotenv
binance
